Add PNIN as a system manifest parameter.
Fixes: #349

Adds a new boolean parameter indicating weather to use PNIN or not to
the system section of the manifest.

The default, False enables tasks for extlinux and grub to pass
net.ifnames=0 biosdevnames=0 parameters to the kernel.

Previously for grub PNIN was disabled for stretch and later. This
remains the default but setting system.pnin True will allow the behavior
to be overridden. The additional parameters have no effect on earlier
releases are passed regardless. Extlinux had no prior special behavior,
it now mirrors grub.